refactor(record): split record.ts into record/ module directory
Split the monolithic src/record.ts (618 lines, 23KB) into 6 focused modules:

- record/types.ts        — shared interfaces (RecordedRequest, RecordResult, RecordOptions)
- record/interceptor.ts  — browser-injected fetch/XHR capture scripts
- record/analysis.ts     — URL pattern, auth detection, array finding, scoring
- record/generator.ts    — YAML candidate generation from recorded requests
- record/session.ts      — session orchestration, polling, analysis output
- record/index.ts        — barrel re-exports for backward-compatible API

Added 36 unit tests covering all extracted pure functions:
- urlToPattern, detectAuthIndicators, findArrayPath
- inferCapabilityName, inferStrategy, scoreRequest
- generateFullCaptureInterceptorJs, generateReadRecordedJs
- buildRecordedYaml

All 259 tests pass (223 existing + 36 new). Zero TypeScript errors.
